rename `parts` folder to visualizers folder in re_space_view_spatial
