+++ crates/re_space_view_spatial/src/contexts/shared_render_builders.rs
@@ -3,7 +3,7 @@ use re_renderer::{LineStripSeriesBuilder, PointCloudBuilder, RenderContext};
 use re_types::ComponentNameSet;
 use re_viewer_context::{IdentifiedViewSystem, ViewContextSystem};
 
-use crate::parts::{
+use crate::visualizers::{
     SIZE_BOOST_IN_POINTS_FOR_LINE_OUTLINES, SIZE_BOOST_IN_POINTS_FOR_POINT_OUTLINES,
 };
