+++ crates/re_space_view_spatial/src/space_view_3d.rs
@@ -11,9 +11,9 @@ use re_viewer_context::{
 use crate::{
     contexts::{register_spatial_contexts, PrimitiveCounter},
     heuristics::{auto_spawn_heuristic, update_object_property_heuristics},
-    parts::{calculate_bounding_box, register_3d_spatial_parts, CamerasPart},
     ui::SpatialSpaceViewState,
     view_kind::SpatialSpaceViewKind,
+    visualizers::{calculate_bounding_box, register_3d_spatial_parts, CamerasPart},
 };
 
 // TODO(andreas): This context is used to determine whether a 2D entity has a valid transform
