+++ crates/re_space_view_spatial/src/visualizers/assets3d.rs
