+++ crates/re_space_view_spatial/src/visualizers/entity_iterator.rs
