+++ crates/re_space_view_spatial/src/visualizers/meshes.rs
