+++ crates/re_space_view_spatial/src/visualizers/transform3d_arrows.rs
